Extracted domain normalization into separate object
Moved the use of the Uri class to normalize the domain string into a separate UriNormalizer class and made DomainParser consume it via an interface
