+++ Nager.PublicSuffix/IDomainNormalizer.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Nager.PublicSuffix
+{
+    public interface IDomainNormalizer
+    {
+        List<string> NormalizeDomainAndExtractParts(string domain, out string normalizedDomain);
+    }
+}
